article, 댓글 작성자 표시, 작성자에게만 글/댓글 삭제 보이도록 설정

+++ articles/views.py
@@ -20,7 +20,9 @@ def detail(request, article_pk):
     # 사용자가 url 에 적어보낸 article_pk 를 통해 디테일 페이지를 보여준다.
     article = get_object_or_404(Article, pk=article_pk)
     comment_form = CommentForm()
-    comments = Comment.objects.all()
+    # comments = Comment.objects.all()  -> 모든 댓글 다 가져옴
+    # comments = article.comment_set.all()  # comment_set -> comments
+    comments = article.comments.all()  # article에 맞는 댓글만 가지고오기
     context = {
         'article': article,
         'comment_form': comment_form,
@@ -37,10 +39,13 @@ def detail(request, article_pk):
 def create(request):
     if request.method == 'POST':
         # Article 을 생성해달라고 하는 요청
-        form = ArticleForm(request.POST)
+        form = ArticleForm(request.POST)  # title, content + user정보도 넣어야한다
         if form.is_valid():
-            form.save()
-            return redirect('articles:index')
+            # form.save() 에서 밑에처럼 바꾼다
+            article = form.save(commit=False)
+            article.user = request.user  # 누가 만든 article인지 얘기해주기
+            article.save()
+            return redirect('articles:detail', article.pk) # article_pk로 하면 오류남  왜??
     else: # GET
         # Article 을 생성하기 위한 페이지를 달라고 하는 요청
         form = ArticleForm()
@@ -51,13 +56,18 @@ def create(request):
 @login_required  # 로그인이 안돼있는 상태에서는 수정할 수 없음
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article_pk)
-    else: # GET method
-        form = ArticleForm(instance=article)
+
+    if article.user == request.user:
+
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article_pk)
+        else: # GET method
+            form = ArticleForm(instance=article)
+    else:
+        return redirect('articles:index')
     context = {'form': form}
     return render(request, 'articles/update.html', context)
 
@@ -68,8 +78,12 @@ def delete(request, article_pk):
     if request.user.is_authenticated:
         # article_pk 에 맞는 article 을 꺼낸다. 삭제한다.
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
-        return redirect('articles:index')
+        # 삭제를 요청하는 user와 article을 작성한 user가 같은 경우에만 삭제
+        if article.user == request.user:
+            article.delete()
+            return redirect('articles:index')
+        else:
+            return redirect('articles:detail', article_pk)
     else:
         return redirect('articles:index')
 
@@ -80,8 +94,10 @@ def comments_create(request, article_pk):
     if request.user.is_authenticated:
         form = CommentForm(request.POST)
         if form.is_valid():
-            comment = form.save(commit=False)
+            comment = form.save(commit=False) # DB에 바로 저장하지 않고 대기시킴
             comment.article_id = article_pk
+            # commet의 user정보를 request.user로 넣겠다
+            comment.user = request.user
             comment.save()
     return redirect('articles:detail', article_pk)
 
@@ -90,6 +106,10 @@ def comments_create(request, article_pk):
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
-        return redirect('articles:detail', article_pk)
+        if request.user == comment.user:
+            comment.delete()
+            return redirect('articles:detail', article_pk)
+        else:
+            return render(request, 'articles/datail', article_pk)
     return HttpResponse('You are Unauthorized', status=401)
+
